move scripts to util file
